Corrected Tag Usage
2 Lines for 1 hr. I don't like this saying as much as 2 birds and 1 stone. 

Look at indentations first!!! Super important with python debugging. 

This allows all tags to be read. The error occurs since in qpaceTagChecker in line 125, the return for the array happens immediately when the first tag is found. This ensures that the returned array does not contain all available tags, and only the first one. This causes problems. 

Correction on qpaceInterpreter ensured that the tags weren't initiated twice.