Tag Checker Correction 2 Queues
Corrected Error where Pi would reserve tags that it used to send to Ground, thus, on occasion the files sent to Pi where considered invalid due to a duplicate tag. This is not the functionality wanted, as the files sent from the WTC should have a Queue that records it's information accurately.

With this fix, only tags used by the Ground Station are recorded in one Queue, and tags sent from the Pi are recorded in another Queue