Began implementing serial console packet transfer
Because of an issue with the WTC, for now we will transfer packets by directly printing them to the console and reading from TX. This push simply stopped the logger from printing if pin 20 on the pi is high. All print statements in the code need to be removed or replaced with a logger call.